streamlit/requirements.txt
streamlit>=1.28.1
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.17.0
google-cloud-bigquery>=3.11.4
google-auth>=2.22.0
google-cloud-geminidataanalytics